Created project structure and made get_windows.py, screen_capture.py, vision_analysis.py
Created a script to get the size and location of the bluestacks window. Created a script using mss to screenshot the bluestacks window. Also created a script using opencv's matchtemplate function. Uses numpy to organize the image for opencv and to find where we are above the threshold in matching the image template in our vision_analysis.py